// src/utils/testId-name.enum.ts
export enum TestIdName {
    AllergensAddBtn = 'add-allergen-button',
    AllergensAddInput = 'add-other-allergen',
    AllergensCheckbox = 'allergen',
    AllergensMenu = 'allergens-menu',
    AllergensMenuBtn = 'allergens-menu-button',
    AllergensMenuBtnFilter = 'allergens-menu-button-filter',
    AllergensSwitcher = 'allergens-switcher',
    AllergensSwitcherFilter = 'allergens-switcher-filter',

    Breadcrumbs = 'breadcrumbs',
    BurgerCloseIcon = 'close-icon',
    BurgerNav = 'nav',
    BurgerOpenIcon = 'hamburger-icon',
    CardLink = 'card-link',

    Carousel = 'carousel',
    CarouselBack = 'carousel-back',
    CarouselForward = 'carousel-forward',

    Checkbox = 'checkbox',

    ErrorNotif = 'error-notification',
    ErrorNotifClose = 'close-alert-button',
    ErrorPageHome = 'error-page-go-home',

    FilterCategory = 'filter-menu-button-категория',
    FilterClearBtn = 'clear-filter-button',
    FilterCloseBtn = 'close-filter-drawer',
    FilterDrawer = 'filter-drawer',
    FilterFindBtn = 'find-recipe-button',
    FilterOpenBtn = 'filter-button',
    FilterTag = 'filter-tag',

    Footer = 'footer',
    Header = 'header',
    IngredientQuantity = 'ingredient-quantity',
    JuiciestLink = 'juiciest-link',
    JuiciestLinkMobile = 'juiciest-link-mobile',
    LoaderApp = 'app-loader',
    LoaderSearch = 'loader-search-block',
    LoadMoreBtn = 'load-more-button',
    MenuActive = 'active',
    SearchBtn = 'search-button',
    SearchInput = 'search-input',
    StepperDecrement = 'decrement-stepper',
    StepperIncrement = 'increment-stepper',
    Tab = 'tab',
    SignInForm = 'sign-in-form',
    InputLogin = 'login-input',
    InputPassword = 'password-input',
    PasswordBtn = 'password-visibility-button',
    SubmitBtn = 'submit-button',
    ForgotPassword = 'forgot-password',
    ModalSignIn = 'sign-in-error-modal',
    ModalClose = 'close-button',
    RepeatBtn = 'repeat-button',
    SignUpProgress = 'sign-up-progress',
    SignUpForm = 'sign-up-form',
    InputFirstName = 'first-name-input',
    InputLastName = 'last-name-input',
    InputEmail = 'email-input',
    InputPasswordConfirm = 'confirm-password-input',
    InputVerificationCode = 'verification-code-input',
    ModalSignUpSuccess = 'sign-up-success-modal',
    ModalEmailVerificationFailed = 'email-verification-failed-modal',
    ModalSendEmail = 'send-email-modal',
    ModalVerificationCode = 'verification-code-modal',
    ModalResetCredentials = 'reset-credentials-modal',
}